clock out problem solved

+++ emp-backend/src/main/java/com/yongyang/Emp_Project/dto/AttendanceDetailDto.java
@@ -14,9 +14,9 @@ public class AttendanceDetailDto {
     private Date attendDate;
     private Date clockIn;
     private Date clockOut;
-    private long totalHour;
-    private long breakHour;
-    private long overTime;
+    private Double totalHour;
+    private Double breakHour;
+    private Double overTime;
     
     
 }
